+++ routes/users_route.js
@@ -3,11 +3,10 @@ const router = express.Router();
 const { Users } = require('../models');
 const authMiddleware = require('../middlewares/auth-middleware.js');
 const upload = require('../middlewares/uploadFile.js');
-const { Console } = require('console');
 
 router.get('/userInfo', authMiddleware, async (req, res) => {
-
   const userid = res.locals.user.userid;
+
   const user = await Users.findOne({
     attributes: ['email', 'profilepicture', 'nickname', 'birth', 'gender', 'address', 'introduce', 'updatedAt'],
     where: { userid },
@@ -28,7 +27,9 @@ router.put('/userInfo', authMiddleware, upload, async (req, res) => {
     res.status(400).json({
       errorMessage: '닉네임은 영문 대소문자와 숫자 3~12자리로 입력해주세요.',
     });
-    fs.unlinkSync('./img-server/' + req.file.filename);
+    if (profilepicture) {
+      fs.unlinkSync('./img-server/' + req.file.filename);
+    }
     return;
   }
 
@@ -36,7 +37,9 @@ router.put('/userInfo', authMiddleware, upload, async (req, res) => {
     res.status(400).json({
       errorMessage: '비밀번호는 영문 대소문자와 특수문자로 이루어진 4~16자리로 입력해주세요.',
     });
-    fs.unlinkSync('./img-server/' + req.file.filename);
+    if (profilepicture) {
+      fs.unlinkSync('./img-server/' + req.file.filename);
+    }
     return;
   }
 
@@ -44,7 +47,9 @@ router.put('/userInfo', authMiddleware, upload, async (req, res) => {
     res.status(400).json({
       errorMessage: '패스워드가 패스워드 확인란과 다릅니다.',
     });
-    fs.unlinkSync('./img-server/' + req.file.filename);
+    if (profilepicture) {
+      fs.unlinkSync('./img-server/' + req.file.filename);
+    }
     return;
   }
 
@@ -55,7 +60,9 @@ router.put('/userInfo', authMiddleware, upload, async (req, res) => {
     res.status(400).json({
       errorMessage: '중복된 닉네임입니다.',
     });
-    fs.unlinkSync('./img-server/' + req.file.filename);
+    if (profilepicture) {
+      fs.unlinkSync('./img-server/' + req.file.filename);
+    }
     return;
   }
 
@@ -63,7 +70,9 @@ router.put('/userInfo', authMiddleware, upload, async (req, res) => {
     res.status(400).json({
       errorMessage: '이미 가입된 이메일입니다.',
     });
-    fs.unlinkSync('./img-server/' + req.file.filename);
+    if (profilepicture) {
+      fs.unlinkSync('./img-server/' + req.file.filename);
+    }
     return;
   }
